refactor(user): 패키지 구조 단순화 및 불필요한 import 제거
- user 모듈의 패키지를 application / domain / presentation / exception 으로 재구성
- dto, mapper, service를 application 하위로 이동
