+++ backend/src/main/java/com/todo/user/application/mapper/UserMapper.java
@@ -1,7 +1,7 @@
-package com.todo.user.mapper;
+package com.todo.user.application.mapper;
 
 import com.todo.user.domain.User;
-import com.todo.user.dto.SignupRequest;
+import com.todo.user.application.dto.SignupRequest;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
