+++ backend/src/test/java/com/todo/cateogry/service/CategoryCommandServiceTest.java
@@ -8,8 +8,7 @@ import com.todo.cateogry.dto.CategoryRequest;
 import com.todo.cateogry.dto.CategoryResponse;
 import com.todo.cateogry.mapper.CategoryMapper;
 import com.todo.user.domain.User;
-import com.todo.user.service.UserCommandService;
-import com.todo.user.service.UserQueryService;
+import com.todo.user.application.service.UserQueryService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
